Fixed the Aloe's footsteps audio source pitch issue
It turns out I fixed the pitch issue in the main function that I use to play sounds, but I forgot to fix it in the separate function I use just for the footstep sounds. This commit rectifies this. The footstep sounds are handled in a different function because its called via an Animation Event on the client side, whereas all the other sounds are called by the server and then played on the clients through a ClientRPC.