Potential fixes for landmine bug and footstep pitch bug
For the landmine bug, I gave the harmony patch that is supposed to prevent the Aloe and kidnapped player from triggering the landmine, a slightly higher than normal priority. I also added a check to make sure that the fake player body ragdoll doesn't trigger the landmine - it should've been there from the beginning, but only just realised now that its needed -_- . For the footstep pitch bug, I think the audio source gets bugged out randomly because I'm changing the pitch of the source multiple times per second. As a potential fix, I added a variable to store the default pitch in the Start() method, and then used that to generate the random pitches, instead of what I was doing previously.