Added a StateCache for the StateManagedAI
Added a StateCache for the StateManagedAI class which will improve the performance of the AI state machine's initialization process. Previously, the `ConstructStateDictionary` function would use reflection to get all `BehaviourState` implementations and their constructors, every single time an AI instance spawned. The StateCache rectifies this issue and makes sure the whole reflection and state dictionary stuff only happens once (per enemy ai type ofc). To put it into perspective, lets imagine we spawn in 5 Aloes and 3 Wax Soldiers. With the old code, it would perform the reflection stuff 8 times (5 + 3), however with the new code, it would do it only twice (once for the Aloe, once for the Wax Soldier). This improvement was long coming, I was just lazy. I also wrapped the Wax Soldier handler logic in comments because its not done yet.