Started making the logic for the pursuit state
The pursuit state is going to use the Strategy pattern, so we can easily replace different search algorithms. This will also be useful to implement in the Aloe's code in the future. The (very basic) search strategy which was added, creates an ordered list of nearby AI nodes based on a score. The score is calculated for each node by using the dot product of the player's (last known, as in before they slipped out of view) velocity, and the direction of the player's (last known) position, to said node. This would be fine for dumb creatures like the Thumper, but for the Wax Soldier it's not anywhere near enough, but a start. The next step would be to upgrade the queue of nodes into a probability map.