Refactored core namespace
The AssetLoading, Attributes, Config and Lang namespaces were moved to Biodiversity.Core (they were in Biodiversity.Util). This was done because these things are core components of the mod, and the Util folder was kinda being used as a dumping ground for new classes that didn't fit in Creatures or Items. The musket now fires and reloads properly (without sounds and visuals yet, however).