Wax soldier can how hold musket ingame
The Wax Soldier now spawns with the musket, and works with the walk, walk sideways, spin, stab and stab idle animations. An attack cooldown was also implemented, along with a timer to tell the Wax Soldier when to give up on his search (this timer will be made more complex in the future, and take into account things like threat levels, wax durability, health, etc). The bayonet however does no damage in the hands of the soldier - that will be implemented next, along with shooting the musket.