Added Imperium insights
The Imperium mod was added to ./Lib, so we can access the methods from it. It can used to add and remove different values to the Imperium insights panel (if the user has the Imperium mod installed in their game). For the Wax Soldier, the Movement Speed and Location insights were removed, and the insights Behaviour State, Speed, Acceleration, Wax Temperature (in degrees celsius) and Wax Durability (from 0 to 100%) were added. Look at the Start() function in WaxSoldierAI.cs to see how this is done. Insights for the Aloe and Musket item were also added.