Added more heat stuff and fixed wax targeting bug
The components HeatEmitter and HeatSensor were added. HeatEmitter is an abstract class that all heat emitters must implement and override the GetHeatRateAt function. That function calculates how much heat that heat emitter contributes at the inputted target world position. The HeatSensor component is used to detect heat, by using a collider to detect HeatEmitters near it. For each near HeatEmitter, their GetHeatRateAt function will be called with the HeatSensor's transform.position as the input. The FireplaceHeatEmitter is the first extension of the HeatEmitter. A bug in the FOV check in BiodiverseAI.HasLineOfSight was fixed, and now the Wax Soldier can target players properly