Wax soldier melt, death and stun logic polished
The stun now works exactly like how it works with the Nutcracker vanilla enemy. The death, reload and both melt animations are fully implemented. In the death animation, the wax soldier lets go of the musket as he is falling, and causes the screen to shake for any nearby players as he slams into the floor. The DirectedConeHeatEmitters are also fully functioning now, the only thing that needs to be changed is moving the soldier's heat sensor origin to his core, instead of his feet.