Fixed bug in the DataStructures.Caching
Renamed the CachedNullable<T> struct to CachedUnityObject<T> where it now forces T to be a unity object. The old CachedNullable struct was using the EqualityComparer (a standard c# reference comparison) instead of using Unity's overloaded == operator. The code for the BulkPopulateDictionary and LazyDictionary (previously called PerKeyCachedDictionary) was simplified. There is also now a different version of the LazyDictionary which is used for Unity objects called LazyUnityObjectDictionary. The ancient monstrosity NullableObject<T> was also finally removed.