Polished the spin and stab attack
The stab attack now has a mini-leap/lunge (no vertical movement, just horizontal) which is in time with the animation. The spin attack now corrects the soldier's rotation to be facing the target player, before and after the spin. As well as making the attack look 'better', it also fixes the issues where the soldier would go into hunting mode instead of pursuit mode after a spin. Fixed the developer scrap config from putting custom settings in its own section, rather than the section where all of the other settings are for that item.